pyproject.toml
[project]
name = "ysearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.4",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
